backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "loss-run-triage"
version = "0.1.0"
description = "Loss Run Triage Agent — AI-powered insurance document processor"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "google-generativeai>=0.8",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "pandas>=2.2",
    "numpy>=1.26",
    "matplotlib>=3.9",
    "weasyprint>=62",
    "jinja2>=3.1",
    "aiofiles>=23.2",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]